Split record-table toolbar controls out of the shared table module.
Moves search, select, chips, dots, and cell clamp/stamp helpers into
record-table-controls.tsx and re-exports them so existing import sites
stay put. Drops the file from the client-route size ratchet.
